# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DJERDJERA Comptable API - Algerian small-business accounting
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /auth/register:
    post:
      tags: [Auth]
      operationId: registerUser
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/login:
    post:
      tags: [Auth]
      operationId: loginUser
      summary: Login and receive access and refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshSession
      summary: Get a new access token using the refresh token cookie
      responses:
        '200':
          description: Refresh successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List fixed assets (equipment) with computed depreciation
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FixedAsset" }
    post:
      operationId: createAsset
      tags: [assets]
      summary: Register a fixed asset (auto-generates monthly amortization log)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FixedAssetInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FixedAsset" }
  /assets/{id}:
    delete:
      operationId: deleteAsset
      tags: [assets]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
  /assets/{id}/amortization:
    get:
      operationId: getAssetAmortization
      tags: [assets]
      summary: Monthly amortization schedule for an asset
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AmortizationEntry" }

  # ============ TRANSACTIONS (JOURNAL) ============
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
        firstName:
          type: string
        lastName:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    TokenResponse:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    FixedAsset:
      type: object
      required: [id, label, costHt, purchaseDate, lifeYears, residualValue, monthlyDepreciation, accumulatedDepreciation, bookValue]
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
        category: { type: string, nullable: true, description: "e.g. Industrial Equipment, Vehicle, Computer" }
        costHt: { type: number }
        purchaseDate: { type: string, format: date }
        lifeYears: { type: integer }
        residualValue: { type: number }
        monthlyDepreciation: { type: number }
        accumulatedDepreciation: { type: number }
        bookValue: { type: number }
    FixedAssetInput:
      type: object
      required: [label, costHt, purchaseDate, lifeYears]
      properties:
        label: { type: string, minLength: 1 }
        category: { type: string, nullable: true }
        costHt: { type: number, minimum: 0 }
        purchaseDate: { type: string, format: date }
        lifeYears: { type: integer, minimum: 1, maximum: 50 }
        residualValue: { type: number, minimum: 0, default: 0 }
    AmortizationEntry:
      type: object
      required: [id, date, amount]
      properties:
        id: { type: string, format: uuid }
        assetId: { type: string, format: uuid }
        date: { type: string, format: date }
        amount: { type: number }
...